Move controllers to a new class
